# Conflicts:
#	src/main/java/web/mvc/controller/StockController.java
#	src/main/java/web/mvc/repository/UserBuyDetailRepository.java
#	src/main/java/web/mvc/service/StockServiceImpl.java
#	src/test/java/web/mvc/StockTest.java

+++ src/main/java/web/mvc/dto/UserChargeDTO.java
@@ -8,8 +8,8 @@ import web.mvc.domain.UserCharge;
 @Getter
 @Setter
 public class UserChargeDTO {
-    private int userChargeSeq;
-    private int managementUserSeq;
+    private Long userChargeSeq;
+    private Long managementUserSeq;
     private String chargeDate;
     private long price;
     private String orderUid;
